refactor and reorganize typing
